feat: fix troubles with headers
