+++ lab_5/ReadMe.md
@@ -1,6 +1,4 @@
-# C хедерами что-то не так
-## Оптимально убрать предварительную декларацию в мэйне заменить на 
-```c
-#include "../headers/fun.h"
-```
-### Fix as soon as possible -_-
+~~# C хедерами что-то не так~~
+
+-[X] Fix as soon as possible -_-
+#### Надо было просто ctr в cpp отдельный для каждого класса всунуть
