+++ lab_5/src/main.cpp
@@ -1,4 +1,4 @@
-void mainMenu();
+#include "../headers/fun.h"
 
 int main() {
     mainMenu();
